battery real rated voltage was already / 100 and back to float

+++ epsolar_commander.h
@@ -80,7 +80,7 @@ extern  float   energyConsumedTotal;
 
 extern  char    controllerClock[];
 
-extern  int     batteryRatedVoltage;        // will be 12, 24,36, 48...
+extern  float   batteryRatedVoltage;        // will be 12, 24,36, 48...
 extern  float   batteryRatedLoadCurrent;
 extern  float   batteryRatedChargingCurrent;
